Gabes commissioning

+++ app/hedge.py
@@ -17,7 +17,7 @@ import subprocess
 import sftp
 from datetime import datetime
 from pathlib import Path
-
+import requests.exceptions
 #  =======================================
 # Object settings
 #  =======================================
@@ -27,7 +27,7 @@ fileRW = output.Output()
 #  =======================================
 # Variable settings
 #  =======================================
-box_id = "box-9082242689124"
+box_id = "box-7373203790301"
 cycle_time = 600
 last_ran = None
 PIZERO_IP = '10.170.1.'
@@ -44,7 +44,7 @@ def read_and_average(measurement_type):
     if measurement_type == "weight":
         weight_sensor = weight.sensor()
         weight_sensor.write("weight.csv", debug=True, iterations=10)  # Read Weight
-        weight_sensor.avrg("weight.csv", "avrgweight.csv", 0.95, True)  # Average Weight
+        weight_sensor.avrg("weight.csv", "avrgweight.csv", True)  # Average Weight
 
     elif measurement_type == "temp":
         thermo_sensor = thermo.sensor()
@@ -71,7 +71,7 @@ def post(box_id, hog_id, to_post):
                     client.create_weight(box_id, "hog-" + hog_id, weight[i], time)
                     fileRW.clear_data("/home/pi/avrgweight.csv")
                 except requests.exceptions.HTTPError as e:
-                    print(getattr(e, "message", repr(e)))
+                    print(getattr(e, "message ", repr(e)))
 
     if to_post["temp"] == True:
         temps_in = fileRW.read("/home/pi/avrgtemp_in.csv", 1)
@@ -91,30 +91,31 @@ def post(box_id, hog_id, to_post):
                 fileRW.clear_data("/home/pi/avrgtemp_out.csv")
 
         except requests.exceptions.HTTPError as e:
-            print(getattr(e, "message", repr(e)))
-
+            print(getattr(e, "message ", repr(e)))
     if to_post["video"] == True:
         os.chdir("/home/pi/Videos")
         files = [glob.glob(e) for e in ["*.mp4"]]
         for file in files[0]:
-            strtime = file.split("_")[0]
-            time = datetime.strptime(strtime, "%Y-%m-%d-%H-%M-%S")
-            try:
-                client.upload_video(
-                    box_id, "hog-" + hog_id, "/home/pi/Videos/" + file, time
-                )
-                os.remove("/home/pi/Videos/" + file)
-            except requests.exceptions.HTTPError as e:
-                print(getattr(e, "message", repr(e)))
+            if file != '1stPass.mp4':
+                strtime = file.split("_")[0]
+                time = datetime.strptime(strtime, "%Y-%m-%d-%H-%M-%S")
+                try:
+                    client.upload_video(
+                        box_id, "hog-" + hog_id, "/home/pi/Videos/" + file, time
+                    )
+                    os.remove("/home/pi/Videos/" + file)
+                except requests.exceptions.HTTPError as e:
+                    print(getattr(e, "message ", repr(e)))
 
 
 def cleanup():
     files_grabbed = [glob.glob(e) for e in ["/home/pi/*.csv"]]
+    print(files_grabbed)  # debuging
     for file in files_grabbed[0]:
         if "avrg" in file:
             pass
         else:
-            fileRW.clear_data(file)
+            print(fileRW.clear_data(file))  # debuging
 
 
 #  =======================================
@@ -123,7 +124,7 @@ def cleanup():
 def main(last_ran, box_id, cycle_time, PIZERO_IP, PIZERO_IP_MIN, PIZERO_IP_MAX, PIZERO_FTP_USERNAME, PIZERO_FTP_PASSWORD):
     start_time = time.time()
     to_post = {"weight": True, "temp": True, "video": True}  # Used for partial posts
-    if pir_sensor.read() == 1:
+    if pir_sensor.read() == 0:  # commissioning change
         print("Started")
         rfid_tag = rfid_sensor.read()[-16:]
         #  =======================================
@@ -150,8 +151,7 @@ def main(last_ran, box_id, cycle_time, PIZERO_IP, PIZERO_IP_MIN, PIZERO_IP_MAX, 
                             "An error has occured, "
                             + i
                             + " will not be posted because"
-                            + getattr(e, "message", repr(e))
-                        )
+                            + getattr(e, "message ", repr(e)))
                         to_post["video"] = False
 
             except Exception as e:
@@ -159,8 +159,7 @@ def main(last_ran, box_id, cycle_time, PIZERO_IP, PIZERO_IP_MIN, PIZERO_IP_MAX, 
                     "An error has occured, "
                     + i
                     + " will not be posted because"
-                    + getattr(e, "message", repr(e))
-                )
+                    + getattr(e, "message ", repr(e)))
                 to_post[i] = False
 
         post(box_id, rfid_tag, to_post)  # Posts data
@@ -184,6 +183,8 @@ def main(last_ran, box_id, cycle_time, PIZERO_IP, PIZERO_IP_MIN, PIZERO_IP_MAX, 
                 sftp.pull_videos(PIZERO_IP + str(i), PIZERO_FTP_USERNAME, PIZERO_FTP_PASSWORD)
                 to_post = {'weight': False, 'temp': False, 'video': True}
                 post(box_id, 'outside', to_post)
+        # weight_sensor = weight.sensor() # Will be run once an hour if PIR not triggered
+        # weight_sensor.tare_weight()  # Commented because awaiting function refactor
         return last_ran
